Rename files and create helper
